requirements.txt
flask
ultralytics
facenet-pytorch
opencv-python
numpy
pillow
torch
torchvision
